requirements.txt
Pillow
pydantic
aiohttp
requests
aiofiles
watchdog
PyGithub
uvicorn[standard]
fastapi
ordered_set
redis